calculates bmi

+++ db/migrate/20170602114031_add_onboarding_attributes_to_user.rb
@@ -7,9 +7,9 @@ class AddOnboardingAttributesToUser < ActiveRecord::Migration
     add_column :users, :town, :string
     add_column :users, :postcode, :string
     add_column :users, :country, :string
-    add_column :users, :height, :integer
+    add_column :users, :height, :float
     add_column :users, :weight, :integer
-    add_column :users, :bmi, :integer
+    add_column :users, :bmi, :float
     add_column :users, :vaccinations, :text
   end
 end
